# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyBusiness 2.0 API",
        "description": "MyBusiness API documentation",
        "contact": {
            "email": "sgg.2020@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/user/index": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all user with roles and permission",
                "operationId": "indexUsers",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user with roles and permission by user id",
                "operationId": "showUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Uuid ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "uuid",
                            "default": "1268A938-37DE-48C6-99C2-99840575104C"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user/add": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "operationId": "bb3f78a92b4c4b3cff80e8dca0c90f52",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "user_name": {
                                                "type": "string"
                                            },
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "middle_name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "panther_id": {
                                                "type": "integer"
                                            },
                                            "avatar": {
                                                "type": "string"
                                            },
                                            "instructor": {
                                                "type": "boolean"
                                            },
                                            "student": {
                                                "type": "boolean"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "active": {
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "department_id": {
                                                "type": "uuid"
                                            },
                                            "roles": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "format": "string",
                                                    "example": [
                                                        "Instructor",
                                                        "Secretary."
                                                    ]
                                                },
                                                "collectionFormat": "multi"
                                            },
                                            "permissions": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "format": "string",
                                                    "example": [
                                                        "Delete a course.",
                                                        "List of program levels."
                                                    ]
                                                },
                                                "collectionFormat": "multi"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Demo",
                                    "user_name": "demo",
                                    "first_name": "first",
                                    "last_name": "last",
                                    "middle_name": "demos",
                                    "email": "demo@fiu.edu",
                                    "panther_id": "6447394",
                                    "avatar": "./../avatar",
                                    "instructor": 1,
                                    "student": 1,
                                    "email_verified_at": "",
                                    "password": "123",
                                    "active": "1",
                                    "created_at": "2024-01-10T22:18:54.927000Z",
                                    "updated_at": "2024-01-10T22:18:54.927000Z",
                                    "department_id": null,
                                    "roles": [
                                        "Instructor",
                                        "Secretary."
                                    ],
                                    "permissions": [
                                        "Delete a course.",
                                        "List of program levels."
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/user/delete/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "String ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "0C6A36BA-10E4-438F-BA86-0D5B68A2BB15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}